Port Http1ClientStageSpec to munit
